Format/type fixes
Ignore-this: 65d18c54d767e637bcc6c63a0dea57ee

darcs-hash:956d754161d49c04d76772dc5d34c7850e35f35c

+++ source/3rdparty/qtcsh/sh.time.c
@@ -734,6 +734,6 @@ pdtimet(clock_t eval, clock_t bval)
     val	= (eval	- bval)	* 100 /	clk_tck;
 #endif /* POSIX	*/
 
-    xprintf("%ld.%02ld", val / 100, val	- (val / 100 * 100));
+    xprintf("%ld.%02ld", (long)(val / 100), (long)(val - (val / 100 * 100)));
 }
 #endif /* BSDTIMES || _SEQUENT_	*/
